Move summarization end-to-end tests into a separate folder
